[ refactor ] Card Modal 中原有的 ModalContent 和 ModalSidebar 拆成獨立元件
