+++ src/components/TrelloCard/Modal/style.ts
@@ -1,7 +1,10 @@
 import styled from "styled-components";
 import { Modal, Layout, Input } from "antd";
 
-const { Sider, Content } = Layout;
+export const ModalStyle: React.CSSProperties = {
+  backgroundColor: "var(--modal-background-color)",
+  color: "var(--modal-color)",
+};
 
 export const TrelloCardModalStyled = styled(Modal)`
   .ant-modal-content {
@@ -22,29 +25,3 @@ export const TitleInputStyled = styled(Input)`
 `;
 
 export const ModalLayoutStyled = styled(Layout)``;
-export const ModalContentStyled = styled(Content)`
-  .contentHeader {
-    display: flex;
-    align-items: center;
-  }
-`;
-export const ModalSidebarStyled = styled(Sider)`
-  .button-link {
-    background-color: var(--modal-button-neutral);
-    border-radius: 3px;
-    color: inherit;
-    cursor: pointer;
-    display: block;
-    height: 32px;
-    margin-top: 8px;
-    max-width: 300px;
-    overflow: hidden;
-    padding: 6px 12px;
-    :hover {
-      background-color: var(--modal-button-neutral-hover);
-    }
-  }
-  .mid {
-    margin-top: 8px;
-  }
-`;
